+++ integration_test/Cargo.toml
@@ -6,6 +6,11 @@ edition = "2018"
 
 [dependencies]
 bitcoincore-rpc = { path = "../client", features = ["rand"] }
-bitcoin = { version = "0.32.0", features = ["serde", "rand"] }
 lazy_static = "1.4.0"
 log = "0.4"
+
+[features]
+default = ["dogecoin"]
+dogecoin = [ "bitcoincore-rpc/dogecoin" ]
+
+
